update delete message functional

+++ src/redux/actions/messages.js
@@ -24,17 +24,19 @@ const Actions = {
     payload: bool,
   }),
   deleteMessageById: (id) => (dispatch) => {
-    messagesApi
-      .deleteById(id)
-      .then(() => {
-        dispatch({
-          type: "MESSAGES:DELETE_MESSAGE",
-          payload: id,
+    if (window.confirm("Ви хочете видалити це повідомлення?")) {
+      messagesApi
+        .deleteById(id)
+        .then(() => {
+          dispatch({
+            type: "MESSAGES:DELETE_MESSAGE",
+            payload: id,
+          });
+        })
+        .catch(() => {
+          dispatch(Actions.setIsLoading(false));
         });
-      })
-      .catch(() => {
-        dispatch(Actions.setIsLoading(false));
-      });
+    }
   },
   fetchMessages: (dialogId) => (dispatch) => {
     dispatch(Actions.setIsLoading(true));
